feat: implement getMovesInDirection utility
